better organization of prompt code
